requirements.txt
python-docx>=1.1.0
requests>=2.28.0
matplotlib>=3.7.0
numpy>=1.24.0
reportlab>=4.0.0
